♻️ Use new TextSmall component
